Updated UIs to remove titles where un-needed also added twitter username
Historically the editing of details was done separatly but as its all done in one now there is no point having the heading as it serves no purpose.

Also this contains an updated twitter api method to get the current user so you can see which account is bound (assuming you have correct scopes).