This should fix scenarios where application is loaded from different dir
This was an issue with `winget` where it would load the application from another directory meaning the cwd would not be where the app was, which was an unexpected scenario.

Also as a minor update have added buttons to browse to the folders to make this easier for users who forget where things are installed.